Remove setupwraith from tvstock
SetupWraith requires read and write gsettings permissions which
are only allowed to platform signed apps. Otherwise the wizard will
crash at the end of the loop before marking the device provisioned,
which causes a plethora of issues. So, ROMs need to ship ship
SetupWraith directly, signing it with their platform key.