Removal of Pixel's Setup Wizard in favor of "Universal" variants (#685)
* Scripts: Remove Data Transfer Tool from gappscore
Using a non-Pixel SetupWizard, we do not need to have this included in installation. The new SetupWizard will not force itself to download the tool either, and usually ending up being stuck in setup after. Notifications (such as the download notification) will also work properly now, instead of having to disable the tool after setup.

* Scripts: Remove libbarhopper library from installation
Not needed anymore with the alternative SetupWizard