Include datatransfertool for pixel devices
When the setupwizard first runs and is "checking for updates" it will
download and install the datatransfertool app if it's a pixel device.
This is a privileged app so it will fail to run properly and cause
setupwizard to crash because the app is not in priv-app dir. So lets
change it to a core package and set conditional to only install on
pixel devices.

Also, this app is the pixel version of the backuprestore app, so set
conditional to not install the backuprestore app on pixel devices.