Add launcherhack to fix pixellauncher overlay issue
The pixel launcher overlay apk gives exclusive special permissions to
the pixel launcher app for the Recents UI. This causes aosp/stock
launchers to crash. The pixel launcher overlay is only installed
in api28 and newer, so this change adds launcher to the removal list
in api28+ to fix this problem.