Improve discovery of the `$TOP` directory
Rather than assume that `$PWD` is also `$TOP`, discover `$TOP` by
`cd`ing to the directory containing the script. This assumes that the
script is not executed via symlink, or, if it is, that the symlink is
in the same directory as the script itself.

Also eliminate the remaining uses of `realpath`.